Stick with the D Kurd 10 for now
Once this works well with the Master the Handpan lessons, should
abstract and generate other scales.

Fix in this preset: the _right_ tonefield should be lower than the
correspond _left_ tonefield; David shows in "1.2 Basic striking and
playing the tone fields" that the pan can be oriented with the two
lowest notes either side of the doum (as in this preset), with the
lowest note on the right.